requirement.txt
flask-expects-json == 1.6.0
flask-orator == 0.2.0
flask == 2.0.2
pyjwt == 2.0.0
python-dotenv == 0.19.1
Flask-Script==2.0.5
gunicorn==20.1.0